Fix the qurt -fPIC issue for good
For qurt, the end target is a shared object so all intermediate static
libraries *must* be compiled with `-fPIC`. To achieve this, we use a
locally defined CMake function `df_add_library` in place of CMake's
`add_library`. The qurt version of this function explicitly forces PIC
whereas the Linux and Darwin versions just pass the parameters directly
to CMake's `add_library`.