Fixed qurt build of df_testapp
The sleep function is missing from latest adsp image so it was added to
the stubs file.

QuRT cannot to a pthread_cond_timedwait on a recursive mutex. It will
reset the DSP. Fixed so the mutex attribute is set to NORMAL except for
NuttX where it is not defined.

Signed-off-by: Mark Charlebois <charlebm@gmail.com>