HRTWorkQueue: use absolute time to get the next scheduling time
This fixes a case where the next scheduling event can be set too high.
It happens if at least two items are processed, and 'next' is calculated
for the first. Then the processing overhead of the second is not considered
and the scheduler sleeps too long. Absolute timestamps avoid this issue.