HRTWorkQueue: avoid scheduling drift (#71)
* HRTWorkQueue: avoid scheduling drift

This makes sure we keep the specified interval over a longer time period.

* HRTWorkQueue: use absolute time to get the next scheduling time

This fixes a case where the next scheduling event can be set too high.
It happens if at least two items are processed, and 'next' is calculated
for the first. Then the processing overhead of the second is not considered
and the scheduler sleeps too long. Absolute timestamps avoid this issue.

* HRTWorkQueue: initialize item with nullptr (just to shut up the RPi compiler)

* framework: remove unused variable