WIP: This breaks DF
The SyncObj unit test was disabled and that masked a but in DF. At some
point the clock was changed to monotonic for Linux but that breaks all
uses of pthread_cond_timedwait which MUST use CLOCK_REALTIME and on Linux
ideally would use CLOCK_REALTIME_RAW.

This PR is trying to resolve what happened to the clocks and to fix them
on Linux. The state on QuRT and NuttX is unknown.

Signed-off-by: Mark Charlebois <charlebm@gmail.com>