LSM9DS1: remove all FIFO logic, fix scaling
The FIFO stuff copied over from the MPU9250 was not really applicable
for this sensor (unless to save energy which we don't need).
I therefore removed all the FIFO logic, and corrected a bunch of other
things along the way. On of them is raising the sample frequency from 67
Hz to 1000 Hz.