MPU9250: correct the interval between samples
It looks like the FIFO doesn't actually deliver at 8 kHz like it is
supposed to but rather at around 6.7 kHz. Therefore, we shouldn't
blindly pass on an interval matching 8 kHz but should adapt it to the
actual performance.

The rate is additionally filtered to lower jitter because of the fact
that we don't always fetch the same number of samples from the FIFO per
_measure() cycle.