DFList: remove lock in get() method
m_sync is there to protect access to the linked list, but the get() methods
do not access the linked list, just items. Therefore the lock is not needed
here (and if this should help prevent accessing deleted objects, the lock
does not help either, since the pointer is passed as argument).