Use time instead of duration for timedwait
This changes the interface of `waitOnSignal` from a time duration for
the timeout to an absolute time (timespec).

This makes more sense because that's how the underlying
`pthread_cond_timedwait` is implemented, and this also gets rid of a
race condition which can occur if the time is updated right after the
time is read using `offsetTime`. This made no difference until now but
with the lockstep SITL implementation this can mean that a new time is
set just before we wait for the timeout. In this case, the timeout will
never trigger because SITL waits for an actuator control message (the
next "lockstep") in order to continue and update the time again.

By making this an absolute time, we don't have this race and would just
return the `pthread_cond_timedwait` straightaway if the time has been
updated in the meantime.